Finalised dialog, improved AdjustPointer
A canvas update is forced when ScrollChoiceBox is initialised so the pointer ends in the right place. The dialog is ready for final release.
Currently need one more fix for an issue with choice selection (can't betray darf until learning floral diplomacy from the bee, even though choice is there it's not selectabnle) and then ready for final release